fix(ph-ai): make creation tools backend native
