+++ ee/hogai/graph/root/tools/create_feature_flag.py
@@ -11,7 +11,6 @@ from posthog.models import FeatureFlag, GroupTypeMapping, Tag, TaggedItem, Team,
 from posthog.models.property_definition import PropertyDefinition
 from posthog.sync import database_sync_to_async
 
-from ee.hogai.graph.taxonomy.agent import TaxonomyAgent
 from ee.hogai.graph.taxonomy.nodes import TaxonomyAgentNode, TaxonomyAgentToolsNode
 from ee.hogai.graph.taxonomy.toolkit import TaxonomyAgentToolkit
 from ee.hogai.graph.taxonomy.tools import TaxonomyTool, ask_user_for_help, base_final_answer
@@ -609,11 +608,7 @@ The tool will automatically:
 
 **Group-based:**
 - "Create a flag targeting organizations"
-- "Create a flag for companies where employee count > 100
-
-**For experiments**: If creating a flag for an A/B test or experiment, after creating
-the flag, you should navigate to the experiments page and use create_experiment with
-this flag's key to complete the experiment setup."
+- "Create a flag for companies where employee count > 100"
     """.strip()
     context_prompt_template: str = "Creates a new feature flag in the project with optional property-based targeting and multivariate variants for A/B testing"
     args_schema: type[BaseModel] = CreateFeatureFlagArgs
